Pulling out the (conceptually optional) factorifyObjects functionality from the serializer. Adding a comment what factoryObjects does anyway.
